+++ src/Http/Kernel.php
@@ -14,7 +14,7 @@ abstract class Kernel extends HttpKernel
      * @var array
      */
     protected $middleware = [
-        \Butler\Service\Http\Middleware\TrustProxies::class,
+        \Illuminate\Http\Middleware\TrustProxies::class,
         \Illuminate\Http\Middleware\HandleCors::class,
         \Illuminate\Foundation\Http\Middleware\PreventRequestsDuringMaintenance::class,
         \Illuminate\Foundation\Http\Middleware\ValidatePostSize::class,