+++ src/ServiceProvider.php
@@ -238,9 +238,11 @@ class ServiceProvider extends BaseServiceProvider
     {
         if ($this->app->runningInConsole()) {
             $configSource = __DIR__ . '/../config/butler.php';
+            $indexSource = __DIR__ . '/../public/index.php';
             $viewsSource = __DIR__ . '/../resources/views';
 
             $this->publishes([$configSource => base_path('config/butler.php')], 'config');
+            $this->publishes([$indexSource => public_path('index.php')], 'butler-index');
             $this->publishes([$viewsSource => resource_path('views/vendor/service')], 'views');
         }
     }