+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Require laravel v9.11.
 
+### Removed
+- Custom TrustProxies middleware.
+
 ## [0.19.0] - 2022-04-28
 
 ### Added