+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.19.1] - 2022-05-20
+
 ### Added
 - Publishable "index.php".
 - Octane listener "ForgetConnections".